Add eLCM APIs to control VIOM table
This patch adds a feature to control VIOM table in iRMC with eLCM API.
VIOM table is updated by setting AdapterConfig profile. This feature
enables virtualizing MACs and WWNs of I/O cards and booting a server
with specifying boot protocol such as PXE, FC and iSCSI. The feature
is required for server redundancy with remote volume.

This patch defines classes representing VIOM table elements and a
client setting this table to iRMC.

VIOM stands for Virtual I/O Manager.

Change-Id: I65369d42d22cfbdea9613a5fe59838cd03754419